+++ creme/optim/__init__.py
@@ -7,7 +7,7 @@ from .ada_delta import AdaDelta
 from .ada_grad import AdaGrad
 from .ada_max import AdaMax
 from .adam import Adam
-from .amsgrad import AMSGrad
+from .ams_grad import AMSGrad
 from .base import Optimizer
 from .ftrl import FTRLProximal
 from .mini_batch import MiniBatcher